performance improvements

+++ word-game-be/model/Message.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type Message struct {
-	ID        *uint32 `gorm:"primarykey" json:"id"`
+	ID        *int64 `gorm:"primarykey" json:"id"`
 	AuthorID  *uint32
 	LobbyID   *int64
 	Message   string
